[スタート待機] refactor :スタート待機モード、出題モードを個別コンポーネントに変更
出題モードコンポーネント( QuestionMode.tsx )
スタート待機モードコンポーネント( WaitingMode.tsx )
